% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_breakfast_data.R
\name{filter_breakfast_data}
\alias{filter_breakfast_data}
\title{Filter the breakfast dataset}
\usage{
filter_breakfast_data(breakfast_df, days_chr, region_chr)
}
\arguments{
\item{breakfast_df}{A data.frame. The breakfast dataset with at least 
the columns `day_of_the_week` and `region`.}

\item{days_chr}{A character vector. The days of week.}

\item{region_chr}{A character vector. The regions to select.}
}
\value{
A data.frame.
}
\description{
Filter the breakfast dataset by days in region.
}
\examples{
filter_breakfast_data(
  breakfast_df = breakfast,
  days_chr = c("Monday", "Tuesday"),
  region_chr = "Center"
)
}
