% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print.cptmodel}
\alias{print.cptmodel}
\title{Change-points estimated by solution path generation + model selection methods}
\usage{
\method{print}{cptmodel}(x, ...)
}
\arguments{
\item{x}{a \code{cptmodel} object}

\item{...}{current not in use}
}
\description{
Print method for objects of class \code{cptmodel}
}
\examples{
f <- rep(rep(c(0, 1), each = 50), 5)
x <- f + rnorm(length(f)) * .5
print(model.ic(sol.idetect(x)))
}
