% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_breakfast.R
\docType{data}
\name{breakfast}
\alias{breakfast}
\title{breakfast}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
  \item{ day_end }{  character } { Weekdays or Weekend }
  \item{ start_time }{  hms,difftime }
  \item{ day_of_the_week }{  character }
  \item{ domicile }{  character }
  \item{ region }{  character }
}
}
\source{
INSEE
}
\usage{
breakfast
}
\description{
A simple dataset about the breakfast time in France..
}
\keyword{datasets}
