% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakfast.R
\docType{package}
\name{breakfast}
\alias{breakfast}
\alias{breakfast-package}
\title{breakfast: Multiple change-point detection and segmentation for data sequences}
\description{
The breakfast package performs multiple change-point detection in data sequences, or sequence
segmentation, using computationally efficient multiscale methods. This version of the
package only implements the "Tail-Greedy Unbalanced Haar" change-point detection and segmentation
methodology; more methods
will be added in future versions of the package. To start with, see the function
\code{segment.mean}.
}
\examples{
#See Examples for segment.mean
}
\references{
"Tail-greedy bottom-up data decompositions and fast multiple change-point detection", P. Fryzlewicz (2017), preprint.
}
\seealso{
\code{\link{segment.mean}}
}
\author{
Piotr Fryzlewicz, \email{p.fryzlewicz@lse.ac.uk}
}
