% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print.breakfast.cpts}
\alias{print.breakfast.cpts}
\title{Change-points estimated by breakfast}
\usage{
\method{print}{breakfast.cpts}(x, ...)
}
\arguments{
\item{x}{a \code{breakfast.cpts} object}

\item{...}{not in use}
}
\description{
Print method for objects of class \code{breakfast.cpts}
}
\examples{
f <- rep(rep(c(0, 1), each = 50), 5)
x <- f + rnorm(length(f))
print(breakfast(x, option = 'user'))
}
